+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## [4.1.0](https://github.com/capacitor-community/app-icon/compare/v4.0.0...v4.1.0) (2023-05-28)
+
+### Features
+
+* **android:** Introducing Android support
+
+## [3.1.0-beta.2](https://github.com/capacitor-community/app-icon/compare/v3.1.0-beta.1...v3.1.0-beta.2) (2023-04-10)
+
+### Features
+
+* **android:** add `isSupported` plugin method
+
 ## [2.0.4](https://github.com/capacitor-community/app-icon/compare/v2.0.3...v2.0.4) (2021-08-31)
 
 ### Minor Changes